Revert "implement ranked moderation (#146)"
This reverts commit 92a06bd8c15345cb07ac372ad0ed44ccf0560006.
