+++ src/main/kotlin/com/chsdngm/tilly/UpdatesPoller.kt
@@ -2,16 +2,13 @@ package com.chsdngm.tilly
 
 import com.chsdngm.tilly.handlers.CommandHandler
 import com.chsdngm.tilly.handlers.MemeHandler
-import com.chsdngm.tilly.handlers.PrivateModerationVoteHandler
 import com.chsdngm.tilly.handlers.VoteHandler
 import com.chsdngm.tilly.model.CommandUpdate
 import com.chsdngm.tilly.model.MemeUpdate
-import com.chsdngm.tilly.model.PrivateVoteUpdate
 import com.chsdngm.tilly.model.VoteUpdate
 import com.chsdngm.tilly.utility.TillyConfig.Companion.BOT_TOKEN
 import com.chsdngm.tilly.utility.TillyConfig.Companion.BOT_USERNAME
 import com.chsdngm.tilly.utility.hasMeme
-import com.chsdngm.tilly.utility.hasPrivateVote
 import com.chsdngm.tilly.utility.hasStatsCommand
 import com.chsdngm.tilly.utility.hasVote
 import org.springframework.stereotype.Component
@@ -21,8 +18,7 @@ import org.telegram.telegrambots.meta.api.objects.Update
 @Component
 class UpdatesPoller(val memeHandler: MemeHandler,
                     val voteHandler: VoteHandler,
-                    val commandHandler: CommandHandler,
-                    val privateModerationVoteHandler: PrivateModerationVoteHandler) : TelegramLongPollingBot() {
+                    val commandHandler: CommandHandler) : TelegramLongPollingBot() {
 
   override fun getBotUsername(): String = BOT_USERNAME
 
@@ -33,7 +29,6 @@ class UpdatesPoller(val memeHandler: MemeHandler,
       update.hasVote() -> voteHandler.handle(VoteUpdate(update))
       update.hasMeme() -> memeHandler.handle(MemeUpdate(update))
       update.hasStatsCommand() -> commandHandler.handle(CommandUpdate(update))
-      update.hasPrivateVote() -> privateModerationVoteHandler.handle(PrivateVoteUpdate(update))
     }
   }
 }
