+++ src/main/kotlin/com/chsdngm/tilly/model/Models.kt
@@ -1,14 +1,91 @@
 package com.chsdngm.tilly.model
 
 import com.chsdngm.tilly.utility.mention
+import org.hibernate.annotations.Type
 import org.telegram.telegrambots.meta.api.objects.Update
 import org.telegram.telegrambots.meta.api.objects.User
 import java.io.File
+import java.io.Serializable
 import java.time.Instant
+import javax.persistence.*
 
-
-private val trashCaptionParts = listOf("sender", "photo from")
 const val week: Long = 60 * 60 * 24 * 7
+private val trashCaptionParts = listOf("sender", "photo from")
+
+@Entity
+data class Meme(
+    @Id val chatMessageId: Int,
+    val senderId: Int,
+    val fileId: String,
+    val caption: String?,
+    val privateMessageId: Int?,
+    val channelMessageId: Int? = null,
+    @OneToMany(cascade = [CascadeType.ALL], fetch = FetchType.LAZY, mappedBy = "chatMessageId", orphanRemoval = true)
+    val votes: MutableList<Vote> = mutableListOf()) {
+
+  override fun toString(): String {
+    return "Meme(chatMessageId=$chatMessageId, senderId=$senderId, caption=$caption, privateMessageId=$privateMessageId, votes=$votes)"
+  }
+}
+
+@Entity
+data class TelegramUser(
+    @Id val id: Int,
+    val username: String?,
+    val firstName: String?,
+    val lastName: String?
+)
+
+@Entity
+@IdClass(Vote.VoteKey::class)
+data class Vote(
+    @Id val chatMessageId: Int,
+    @Id val voterId: Int,
+    @Enumerated(EnumType.STRING) val value: VoteValue,
+    @Enumerated(EnumType.STRING) val source: VoteSourceType
+) {
+
+  @Embeddable
+  data class VoteKey(
+      val chatMessageId: Int,
+      val voterId: Int) : Serializable
+}
+
+@Entity
+data class Image(
+    @Id val fileId: String,
+    @Lob @Type(type = "org.hibernate.type.BinaryType") val file: ByteArray
+) {
+
+  override fun equals(other: Any?): Boolean {
+    if (this === other) return true
+    if (javaClass != other?.javaClass) return false
+
+    other as Image
+
+    if (fileId != other.fileId) return false
+    if (!file.contentEquals(other.file)) return false
+
+    return true
+  }
+
+  override fun hashCode(): Int {
+    var result = fileId.hashCode()
+    result = 31 * result + file.contentHashCode()
+    return result
+  }
+}
+
+
+enum class VoteValue(val emoji: String) {
+  UP("\uD83D\uDC8E"),
+  DOWN("\uD83D\uDCA9")
+}
+
+enum class VoteSourceType {
+  CHAT,
+  CHANNEL
+}
 
 class VoteUpdate(update: Update) {
   val fromId: Int = update.callbackQuery.from.id
@@ -27,7 +104,7 @@ class VoteUpdate(update: Update) {
   }
 }
 
-class MemeUpdate(update: Update) {
+open class MemeUpdate(update: Update) {
   val messageId: Int = update.message.messageId
   val caption: String? = update.message.caption?.takeIf { caption ->
     val lowerCaseCaption = caption.toLowerCase()
@@ -63,10 +140,4 @@ class CommandUpdate(update: Update) {
   override fun toString(): String {
     return "CommandUpdate(senderId=$senderId, value=$value)"
   }
-}
-
-class PrivateVoteUpdate(update: Update) {
-  val senderId: Long = update.callbackQuery.message.chatId
-  val messageId: Int = update.callbackQuery.message.messageId
-  val voteValue: PrivateVoteValue = PrivateVoteValue.valueOf(update.callbackQuery.data)
 }
