--- src/main/kotlin/com/chsdngm/tilly/model/enums.kt
@@ -1,17 +1,0 @@
-package com.chsdngm.tilly.model
-
-enum class VoteValue(val emoji: String) {
-  UP("\uD83D\uDC8E"),
-  DOWN("\uD83D\uDCA9")
-}
-
-enum class VoteSourceType {
-  CHAT,
-  CHANNEL,
-  PRIVATE_CHAT
-}
-
-enum class PrivateVoteValue(val emoji: String) {
-  APPROVE("\uD83D\uDC8E"),
-  DECLINE("\uD83D\uDCA9")
-}
